video workers moved to a directory
